Refactor and organize code into packages (A-Packages)
