--- src/main/java/InvalidArgumentException.java
@@ -1,5 +1,0 @@
-public class InvalidArgumentException extends DukeException {
-    public InvalidArgumentException() {
-        super("Hmm, I recognized your command but it seems like the arguments are not complete / invalid.");
-    }
-}
