--- src/main/java/InvalidCommandException.java
@@ -1,5 +1,0 @@
-public class InvalidCommandException extends DukeException {
-    public InvalidCommandException() {
-        super("Sorry, I did not recognize your command.");
-    }
-}
