+++ src/main/java/duke/Ui.java
@@ -1,3 +1,7 @@
+package duke;
+
+import duke.util.Constants;
+
 import java.util.List;
 import java.util.Arrays;
