+++ src/main/java/duke/exception/InvalidIndexException.java
@@ -1,3 +1,5 @@
+package duke.exception;
+
 public class InvalidIndexException extends DukeException {
     public InvalidIndexException() {
         super("Sorry, your argument has invalid Index. Hint: Use \"list\" command to help you");
