+++ src/main/java/duke/util/Constants.java
@@ -1,3 +1,5 @@
+package duke.util;
+
 public class Constants {
     public static final String LINE = "___________________________________________________________________________";
     public static final String GAP = "      ";
